New client storing and handling system on the server-side
[Change]
On server-side new clients are now stored in a list for those clients
that didn't send any info (id, name).

When a client sends its info it is moved to a hash dictionary with key
being its id and value being pointer to the client.

[New]
Server now has updateClientInfo() method.

[Fix]
Fixed blockReceived() signal not being connected to the slot in chat
extension